clinic/appointment/serializers.py
from rest_framework import serializers
from .models import Appointment
from core.models import Doctor, Patient


class AppointmentSerializer(serializers.ModelSerializer):
    doctor = serializers.PrimaryKeyRelatedField(queryset=Doctor.objects.all())
    patient = serializers.PrimaryKeyRelatedField(queryset=Patient.objects.all())

    class Meta:
        model = Appointment
        fields = '__all__'
