clinic/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.4.2
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
